requirements-dev.txt
allure-pytest==2.8.19
allure-python-commons==2.8.19
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
bleach==3.2.1
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
colorama==0.4.4
coverage==5.3
distlib==0.3.1
docutils==0.16
filelock==3.0.12
identify==1.5.6
idna==2.10
iniconfig==1.1.1
isort==5.6.4
keyring==21.4.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodeenv==1.5.0
packaging==20.4
pkginfo==1.6.1
pluggy==0.13.1
pre-commit==2.8.2
py==1.9.0
pycodestyle==2.6.0
Pygments==2.7.2
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-sugar==0.9.4
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==28.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.51.0
twine==3.2.0
urllib3==1.25.11
virtualenv==20.1.0
webencodings==0.5.1
wrapt==1.12.1
